# Improve search performance with PostgreSQL