--- a/src/metabase/models/card.clj
+++ b/src/metabase/models/card.clj
@@ -281,7 +281,6 @@
             (field-values/update-field-values-for-on-demand-dbs! newly-added-param-field-ids)))))
     ;; make sure this Card doesn't have circular source query references if we're updating the query
     (when (:dataset_query changes)
-      ;; TODO change dataset_query_tokens for postgres
       (check-for-circular-source-query-references changes))
     ;; Make sure any native query template tags match the DB in the query.
     (check-field-filter-fields-are-from-correct-database changes)