--- a/src/metabase/search/scoring.clj
+++ b/src/metabase/search/scoring.clj
@@ -17,10 +17,11 @@
   (str/lower-case query))
 
 (s/defn tokenize :- [s/Str]
-  "Break a search `query` into its constituent tokens"
+  "Break a search `query` into its constituent tokens
+   Note 'foo.bar' is tokenized to ['foo.bar'] but 'foo(bar)' is tokenized to ['foo', 'bar']"
   [query :- s/Str]
   (filter seq
-          (str/split query #"\s+")))
+          (str/split query #"[\s+\(\)\[\]\{\}]")))
 
 (def ^:private largest-common-subseq-length
   (memoize/fifo