This PR speeds up `/api/search` for customers using Postgres for their Metabase DB. It fixes the search performance issues reported here https://github.com/metabase/metabase/issues/17631.

Previously, a search would involve a full table scan on all searchable columns of `report_card` (except those searches excluding the card model). The problem is we were performing substring matches on native SQL with `dataset_query LIKE '%<search-word>%'`, which doesn't scale well across a lot of long SQL queries. The issue was introduced in 39 when we introduced [searches on native SQL](https://github.com/metabase/metabase/pull/14583).

**Summary of changes**
We can use the [full-text search features built into Postgres](https://www.postgresql.org/docs/current/textsearch.html) to replace substring matches on native SQL queries. Unlike the `pg_trgm` extension this requires no extra setup for the user.

Only on Postgres, an extra column will be added to the report_card table, `dataset_query_tokens` of type `tsvector`. This column contains an index of tokens in the `dataset_query` SQL string. Every time a native query card is saved, this value will update.

Instead of searching with `dataset_query LIKE '%<search-word>%'`, we can now use `dataset_query_tokens @@ to_tsquery('<search-word>')`, which is much faster.

**Potential problems?**
1. We might not allow different columns by DB? As far as I know this could be the first time different DBs have needed different schemas.
2. The initial migration could take a long time, because the index must be created for each native SQL query. With 100,000 native SQL questions, the initial migration took 5-10 minutes. This should only affect customers which are having the problem in the first place though, so they should be happy to wait.

**Benchmarking**
With a realistic-ish mix of 100,000 native SQL questions, I found a consistent ~6x speedup in response time from the BE across a range of search strings. For perspective, the longest search I tested improved from 26s to 4s, and the shortest from 7.5s to 1.2s.

**Changes in behaviour**
There some minor differences in the behaviour of full text search, compared to previous search behaviour. A full text search column is essentially an index of words, so substrings of words won't match. Unfortunately the default parser from Postgres parses strings with a dot (`.`) as one word, so `b.bar` would be indexed as one word in `select foo(b.bar) from baz b`. This is customizable and could be avoided, but doing so would add steps to setting up a Metabase instance.

To be clear, here's a summary of differences before and after the change:

Native SQL of question | Search query | before | after | explanation
-- | -- | -- | -- | --
select foo(b.bar) from baz b | ba | Y | N | ba isn't indexed
select foo(b.bar) from baz b | bar | Y | N | bar isn't indexed
select foo(b.bar) from baz b | b.bar | Y | Y | b.bar is indexed


TODO
- [ ] Adjust for Postgres 9.4
- [ ] Write tests
- [x] Benchmarking

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/metabase/metabase/24171)
<!-- Reviewable:end -->