@howonlee 2022-07-29
Note that this would be the first sproc of any kind in MB migrations so you should probably gather wider eng assent for this - iirc we rejected this previously because the previous suggested time we considered sprocs, it was pretty domain-specific to mysql JSON functions (see https://app.slack.com/client/T078VCLCR/CKZEMT1MJ internal slack)